chore: added deployments page container and appropriate routes
