+++ src/ui/layouts/listing-page.tsx
@@ -2,9 +2,10 @@ import { ApplicationSidebar } from "../shared";
 
 type Props = {
   children: React.ReactNode;
+  margin?: boolean;
 };
 
-export function ListingPageLayout({ children }: Props) {
+export function ListingPageLayout({ children, margin = true }: Props) {
   return (
     <>
       <div>
@@ -14,8 +15,8 @@ export function ListingPageLayout({ children }: Props) {
 
         <div className="md:pl-64 flex flex-col flex-1">
           <main className="flex-1">
-            <div className="mx-auto px-4 sm:px-6 md:px-8">
-              <div className="py-4">{children}</div>
+            <div className={margin ? "mx-auto px-4 sm:px-6 md:px-8" : ""}>
+              <div className={margin ? "py-4" : ""}>{children}</div>
             </div>
           </main>
         </div>
